+++ pkg/apis/platform/v1alpha1/zz_generated.deepcopy.go
@@ -100,6 +100,11 @@ func (in *ArangoPlatformChartSpec) DeepCopyInto(out *ArangoPlatformChartSpec) {
 		*out = make(v1.Data, len(*in))
 		copy(*out, *in)
 	}
+	if in.Overrides != nil {
+		in, out := &in.Overrides, &out.Overrides
+		*out = make(v1.Any, len(*in))
+		copy(*out, *in)
+	}
 	return
 }
 
@@ -127,6 +132,11 @@ func (in *ArangoPlatformChartStatus) DeepCopyInto(out *ArangoPlatformChartStatus
 		for i := range *in {
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
+	}
+	if in.Overrides != nil {
+		in, out := &in.Overrides, &out.Overrides
+		*out = make(v1.Any, len(*in))
+		copy(*out, *in)
 	}
 	return
 }